+++ CHANGELOG.md
@@ -3,7 +3,12 @@ Changelog
 
 # Unreleased
 
+# 11.2.0
+
 - `Element.className` is always a string
+- Better support for HTMLSelectElement
+- Expose options and selectedOptions on select element as live lists
+- Refactor NodeList and its descendents, e.g. HTMLCollection and RadioNodeList
 
 # 11.1.0